import-politicians/requirements.txt
# Dependencies for Politics Trust French Data Integration
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Supabase client
supabase>=2.0.0

# Date and time handling
python-dateutil>=2.8.0

# Logging and monitoring (optional but recommended)
colorlog>=6.7.0

# Data validation (optional)
pydantic>=2.0.0

# HTTP session management
requests-cache>=1.1.0

# Progress bars for long operations (optional)
tqdm>=4.65.0

# JSON handling improvements
ujson>=5.8.0

# CSV handling improvements
chardet>=5.2.0